refactor(#741): refactor documentController
